Set status to refunded right after refund process success.
This is because Midtrans sandbox been acting weird by not sending any notifications to our api (demo2), so we can't rely on it to update our internal trx status. By setting it manually after refund request, we make sure customer see the proper status in the trx detail page. This commit also introduces new queue/job to handle the refund data and store it on our database.